test: isolate target rejection fixture
